feat(website): host-aware / + move catalog under /leaderboards
Phase 2 of the FlawlessNES website work.

URL hierarchy:
- / on flawlessnes.com → marketing landing page (Hero, Features,
  Creators).
- / on leaderboards.retrochallenges.com → still serves the catalog
  via host-based routing (headers().get('host')) so the legacy URL
  doesn't break for current users.
- /leaderboards → canonical catalog URL (game-tile grid + recent
  activity feed). Same component the legacy host renders at /.
- /leaderboards/g/[game] and /leaderboards/c/[game]/[challenge] —
  moved from /g/* and /c/* so individual leaderboards live under
  the catalog hierarchy. URL helpers (gameHref, challengeHref) and
  breadcrumbs updated; the API search-index inherits the new URLs
  for free since it builds them through those helpers.
- /u/[userId] stays at root (user profiles are person-scoped, not
  leaderboard-scoped — analog of github.com/<user>).

LEGACY_CATALOG_HOSTS is a single-element set today; when we cut
over fully, drop the host check + add a 301 redirect from the
legacy host to flawlessnes.com.

Catalog rendering extracted into src/components/CatalogView.tsx so
both the legacy / and the new /leaderboards routes render the same
server component.

Landing page is intentionally minimal for now — placeholder Hero
with download + leaderboards CTAs, three feature blurbs, Twitch
creators (slackanater first per app-side ordering). Full polish +
sign-in lands in Phase 3.

Tests updated for new helper outputs; build verified clean.

Also fixes a stray backslash href ("\") in the game-detail
breadcrumb that snuck in during Phase 1.
